--- a/core/Cargo.toml
+++ b/core/Cargo.toml
@@ -17,7 +17,8 @@ doctest = false
 
 [features]
 default = ["std"]
-std = ["math/std", "winter-utils/std", "miden-crypto/std"]
+std = ["miden-crypto/std", "math/std", "winter-utils/std"]
+sve = ["miden-crypto/sve", "std"]
 
 [dependencies]
 math = { package = "winter-math", version = "0.6", default-features = false }