--- a/core/Cargo.toml
+++ b/core/Cargo.toml
@@ -22,7 +22,7 @@ sve = ["miden-crypto/sve", "std"]
 
 [dependencies]
 math = { package = "winter-math", version = "0.6", default-features = false }
-miden-crypto = {git = "https://github.com/0xPolygonMiden/crypto", branch = "next", default-features = false }
+miden-crypto = { package = "miden-crypto", version = "0.7", default-features = false }
 winter-crypto = { package = "winter-crypto", version = "0.6", default-features = false }
 winter-utils = { package = "winter-utils", version = "0.6", default-features = false }