--- a/miden/Cargo.toml
+++ b/miden/Cargo.toml
@@ -41,6 +41,7 @@ concurrent = ["prover/concurrent", "std"]
 default = ["std"]
 executable = ["dep:env_logger", "dep:hex", "hex?/std", "std", "dep:serde", "serde?/std", "dep:serde_derive", "dep:serde_json", "serde_json?/std", "dep:clap", "dep:rustyline"]
 std = ["assembly/std", "log/std", "processor/std", "prover/std", "verifier/std"]
+sve = ["processor/sve", "prover/sve", "std"]
 
 [dependencies]
 assembly = { package = "miden-assembly", path = "../assembly", version = "0.7", default-features = false }