--- a/processor/Cargo.toml
+++ b/processor/Cargo.toml
@@ -9,7 +9,7 @@ repository = "https://github.com/0xPolygonMiden/miden-vm"
 categories = ["emulators", "no-std"]
 keywords = ["miden", "virtual-machine"]
 edition = "2021"
-rust-version = "1.67"
+rust-version = "1.73"
 
 [lib]
 bench = false
@@ -22,7 +22,7 @@ internals = []
 std = ["vm-core/std", "winter-prover/std", "log/std"]
 
 [dependencies]
-log = "0.4.14"
+log = "0.4"
 vm-core = { package = "miden-core", path = "../core", version = "0.7", default-features = false }
 miden-air = { package = "miden-air", path = "../air", version = "0.7", default-features = false }
 winter-prover = { package = "winter-prover", version = "0.6", default-features = false }
@@ -30,7 +30,6 @@ winter-prover = { package = "winter-prover", version = "0.6", default-features =
 [dev-dependencies]
 logtest = { version = "2.0", default-features = false  }
 miden-assembly = { package = "miden-assembly", path = "../assembly", version = "0.7", default-features = false }
-rand-utils = { package = "winter-rand-utils", version = "0.6" }
 test-utils = { package = "miden-test-utils", path = "../test-utils" }
 winter-fri = { package = "winter-fri", version = "0.6" }
 winter-utils = { package = "winter-utils", version = "0.6" }