--- a/processor/src/operations/crypto_ops.rs
+++ b/processor/src/operations/crypto_ops.rs
@@ -179,7 +179,7 @@ mod tests {
         Process,
     };
     use crate::{AdviceInputs, StackInputs, Word, ZERO};
-    use rand_utils::rand_vector;
+    use test_utils::rand::rand_vector;
     use vm_core::{
         chiplets::hasher::{apply_permutation, STATE_WIDTH},
         crypto::merkle::{MerkleStore, MerkleTree, NodeIndex},