--- a/processor/src/operations/ext2_ops.rs
+++ b/processor/src/operations/ext2_ops.rs
@@ -38,7 +38,7 @@ mod tests {
         Process,
     };
     use crate::{StackInputs, ZERO};
-    use rand_utils::rand_value;
+    use test_utils::rand::rand_value;
     use vm_core::QuadExtension;
 
     // ARITHMETIC OPERATIONS