--- a/processor/src/operations/fri_ops.rs
+++ b/processor/src/operations/fri_ops.rs
@@ -243,7 +243,7 @@ mod tests {
     use super::{
         ExtensionOf, Felt, FieldElement, Operation, Process, QuadFelt, StarkField, TWO, TWO_INV,
     };
-    use rand_utils::{rand_array, rand_value, rand_vector};
+    use test_utils::rand::{rand_array, rand_value, rand_vector};
     use vm_core::{utils::collections::Vec, StackInputs};
     use winter_prover::math::{fft, get_power_series_with_offset};
     use winter_utils::transpose_slice;