--- a/prover/Cargo.toml
+++ b/prover/Cargo.toml
@@ -9,7 +9,7 @@ repository = "https://github.com/0xPolygonMiden/miden-vm"
 categories = ["cryptography", "emulators", "no-std"]
 keywords = ["miden", "prover", "stark", "zkp"]
 edition = "2021"
-rust-version = "1.67"
+rust-version = "1.73"
 
 [features]
 concurrent = ["processor/concurrent", "std", "winter-prover/concurrent"]
@@ -25,5 +25,5 @@ winter-prover = { package = "winter-prover", version = "0.6", default-features =
 
 [target.'cfg(all(target_arch = "aarch64", target_os = "macos"))'.dependencies]
 ministark-gpu = { version = "0.1", features = [ "winterfell" ], optional = true }
-elsa = { version = "1.7", optional = true }
+elsa = { version = "1.9", optional = true }
 pollster = { version = "0.3", optional = true }