--- a/test-utils/Cargo.toml
+++ b/test-utils/Cargo.toml
@@ -9,7 +9,7 @@ repository = "https://github.com/0xPolygonMiden/miden-vm"
 categories = ["development-tools::testing", "no-std"]
 keywords = ["miden", "test", "virtual-machine"]
 edition = "2021"
-rust-version = "1.67"
+rust-version = "1.73"
 
 [features]
 default = ["std"]
@@ -19,11 +19,11 @@ std = ["assembly/std", "processor/std", "prover/std", "verifier/std", "vm-core/s
 assembly = { package = "miden-assembly", path = "../assembly", version = "0.7", default-features = false }
 processor = { package = "miden-processor", path = "../processor", version = "0.7", features = ["internals"], default-features = false }
 prover = { package = "miden-prover", path = "../prover", version = "0.7", default-features = false }
-test-case = "3.0.0"
+test-case = "3.2"
 verifier = { package = "miden-verifier", path = "../verifier", version = "0.7", default-features = false }
 vm-core = { package = "miden-core", path = "../core", version = "0.7", default-features = false }
 winter-prover = { package = "winter-prover", version = "0.6", default-features = false }
 
 [target.'cfg(not(target_family = "wasm"))'.dependencies]
-proptest = { version = "1.1"  }
+proptest = { version = "1.3"  }
 rand-utils = { package = "winter-rand-utils", version = "0.6" }