This PR adds Graviton compilation target to the makefile. When compiling to this target we enable `sve` feature on `miden-crypto` recently introduced in https://github.com/0xPolygonMiden/crypto/pull/190.

I also bumped up minimum `rustc` version to 1.73 (this should address #917) and generally cleaned up `.toml` files a bit.

### Performance impact
The time it takes to generate a proof for a program taking $2^{20}$ cycles on a 64-core Graviton3 instance:

| Hash function used | SVE acceleration | Time (seconds)  | Implied frequency | Baseline multiple |
| ------------------ | :--------------: | :-------------: | :---------------: | :---------------: |
| BLAKE3             | No               | 3.9             | 268 KHz           | 1x                |
| RPO256             | No               | 22.1            | 47 Khz            | 5.7x              |
| RPO256             | Yes              | 15.2            | 69 Khz            | 3.9x              |